Completed initial phase of refactoring Interpreter for runtime Polymorphism.
