+++ source/functional_tests.cpp
@@ -39,7 +39,7 @@ class OpcodeTest : public FunctionalTest {
 };
 
 int main() {
-  StandardInterpreter* interpreter = new StandardInterpreter();
+  Interpreter* interpreter = Interpreter::create(Interpreter::TYPE_DEBUGGING);
   if (interpreter) {
     OpcodeTest::dummy(interpreter);
