+++ source/op_vector.cpp
@@ -14,6 +14,8 @@
 #include "include/vm_core.hpp"
 #include "include/vm_inline.hpp"
 #include "include/machine.hpp"
+#include "include/vm_interpreter_standard.hpp"
+#include "include/vm_interpreter_debugging.hpp"
 #include <cstdio>
 
 #include <limits.h>
