+++ source/vm_core.cpp
@@ -25,11 +25,12 @@
 #endif
 
 using std::nothrow;
-
 using namespace ExVM;
 
-
-
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// Base Interpreter
+//
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 Interpreter::Interpreter(size_t rStackSize, size_t dStackSize, size_t cStackSize) :
@@ -66,11 +67,6 @@ Interpreter::~Interpreter() {
   delete[] regStackBase;
 }
 
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void StandardInterpreter::dump() {
-
-}
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -87,6 +83,36 @@ void Interpreter::setCodeSymbolTable(uint16** symbol, uint32 count) {
 void Interpreter::setDataSymbolTable(void** symbol, uint32 count) {
   dataSymbol      = symbol;
   dataSymbolCount = count;
+}
+
+
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// StandardInterpreter
+//
+// Represents the core interpreter for the virtual machine and is primarily responsible for the runtime execution of
+// virtual code.
+//
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#include "include/vm_interpreter_standard.hpp"
+
+StandardInterpreter::StandardInterpreter(size_t rStackSize, size_t dStackSize, size_t cStackSize) :
+Interpreter(rStackSize, dStackSize, cStackSize) {
+
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+StandardInterpreter::~StandardInterpreter() {
+
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+void StandardInterpreter::dump() {
+
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -110,17 +136,12 @@ void StandardInterpreter::execute() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-StandardInterpreter::StandardInterpreter(size_t rStackSize, size_t dStackSize, size_t cStackSize) :
-Interpreter(rStackSize, dStackSize, cStackSize) {
-
-}
-
-StandardInterpreter::~StandardInterpreter() {
-
-}
-
+//
+// DebuggingInterpreter
+//
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#include "include/vm_interpreter_debugging.hpp"
 
 const char* DebuggingInterpreter::statusCodes[] = {
   "Running",
@@ -229,5 +250,16 @@ void DebuggingInterpreter::execute() {
     totalTime, nativeTime, virtualTime, mips
   );
 
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+Interpreter* Interpreter::create(Interpreter::Type type, size_t rStackSize, size_t dStackSize, size_t cStackSize) {
+  switch (type) {
+    case TYPE_DEBUGGING:
+      return new(nothrow) DebuggingInterpreter(rStackSize, dStackSize, cStackSize);
+    default:
+      return new(nothrow) StandardInterpreter(rStackSize, dStackSize, cStackSize);
+  }
 }
